requirements.txt
requests>=2.31.0
rich>=13.7.0
pyyaml>=6.0.1
keyring>=24.3.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-cov>=4.1.0
mypy>=1.7.0